requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==44.0.2
datetime-truncate==1.1.1
diff-match-patch==20241021
Django==5.1.8
django-admin-charts==1.5.0
django-bower==5.2.0
django-cachalot==2.7.0
django-celery-beat==2.8.1
django-debug-toolbar==5.1.0
django-import-export==4.3.7
django-json-widget==2.0.1
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-redis==6.0.0
django-rest-framework==0.1.0
django-simpleui==2025.1.13
django-timezone-field==7.1
django-user-agents==0.4.0
django-silk==5.4.3
django_celery_results==2.6.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
gunicorn==23.0.0
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
pycparser==2.22
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
